requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==4.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.3.0
PyJWT==2.6.0
python-magic==0.4.27
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
